% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLP.R
\name{MLP}
\alias{MLP}
\title{MLP}
\usage{
MLP(input, profiler = "motus25", output = "load")
}
\arguments{
\item{input}{Species-level (metagenome) or genus-level (16S rRNA) taxonomic profile of the gut microbiome.}

\item{profiler}{A profiler and its version used to generate the input specie-level taxonomic profile. "motus25", "motus3", "metaphlan3", "metaphlan4", or "rdp_train_set_16 (16S rRNA)".}

\item{output}{Output format. "load": predicted microbial loads, or "qmp": taxonomic profile that takes into account the predicted loads.}
}
\description{
Function to predict fecal microbial load based on species-level taxonomic profile of the human gut microbiome.
This function accepts the defaults outputs by mOTUs (v2.5 and v3), metaphlan (v3 and v4) profilers or .
}
