% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLpredictor.R
\name{MLpredictor}
\alias{MLpredictor}
\title{MLpredictor}
\usage{
MLpredictor(d, profiler = "motus2", output = "load")
}
\arguments{
\item{d}{Species-level taxonomic profile of the gut microbiome.}

\item{profiler}{A profiler and its version used to generate the specie-level taxonomic profile. "motus2", "motus3", "metaphlan3", or "metaphlan4".}

\item{output}{Output format. "load": predicted microbial loads, or "qmp": taxonomic profile that takes into account the predicted loads.}
}
\description{
Function to predict fecal microbial load based on species-level taxonomic profile of the human gut microbiome.
This function accepts the defaults outputs by mOTUs (v2.5 and v3) and metaphlan (v3 and v4) profilers.
}
